Doubly linked lists

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -3,14 +3,13 @@
 /**
  * print_dlistint - Prints all the elem
  * of a dlistint_t list
- * @dlistint_t: list name
  * @h: pointer
  * Return: num of nodes
  */
 
 size_t print_dlistint(const dlistint_t *h)
 {
-	size_t node_cnt = 0;
+	int node_cnt = 0;
 
 	if (h == NULL)
 		return (node_cnt);
